pyproject.toml
[project]
name = "ragyt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.2",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.22",
    "langchain-mistralai>=1.1.2",
    "langchain-pinecone>=0.2.13",
    "langchain-text-splitters>=1.1.1",
    "mistralai>=2.1.3",
    "pinecone>=7.3.0",
    "pymongo[srv]>=4.16.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
    "youtube-transcript-api>=1.2.4",
]
